Standa: Fix a VC++ warning.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@11065 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/Standa/Standa.h
@@ -73,7 +73,7 @@ private:
    long curSteps_;
    double stepSizeUm_;
    double axisLimitUm_;
-   int stepDivisor_;
+   BYTE stepDivisor_;
    float stageSpeed_;
    MM::TimeoutMs* timeOutTimer_;
    std::string deviceString_;
